Add database schema docs
