--- common/bitcoin-ledger-signer/src/interfaces/BitcoinLedgerSignerOptions.ts
@@ -1,5 +1,0 @@
-export interface BitcoinLedgerSignerOptions {
-    type?: string
-    path?: string
-    baseURL?: string
-}
