+++ common/data/package.json
@@ -6,13 +6,13 @@
         "build": "echo '@casimir/data build not specified. Disregard this warning and any listed errors above if @casimir/types is not needed for the current project build.' && exit 0",
         "clean": "npx rimraf ./scripts/pgdata ./scripts/sql",
         "configure:python": "poetry install && poetry run ipython kernel install --user --name=casimir-data",
-        "dev": "ts-node --transpile-only scripts/db.ts --clean \"$npm_config_clean\" --tables \"$npm_config_tables\"",
-        "seed": "ts-node --transpile-only scripts/seed.ts",
-        "watch": "ts-node-dev --watch src --respawn --transpile-only scripts/db.ts --clean \"$npm_config_clean\" --tables \"$npm_config_tables\"",
+        "dev:postgres": "ts-node --transpile-only scripts/postgres.ts --clean \"$npm_config_clean\" --tables \"$npm_config_tables\"",
+        "watch:postgres": "ts-node-dev --watch src --respawn --transpile-only scripts/postgres.ts --clean \"$npm_config_clean\" --tables \"$npm_config_tables\"",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "dependencies": {
-        "@aws-cdk/aws-glue-alpha": "^2.33.0-alpha.0"
+        "@aws-cdk/aws-glue-alpha": "^2.33.0-alpha.0",
+        "pg": "^8.10.0"
     },
     "devDependencies": {
         "@types/node": "^17.0.38",
