--- common/ethers-helpers/tsconfig.json
@@ -1,24 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ESNext",
-        "strict": true,
-        "preserveConstEnums": true,
-        "noEmit": true,
-        "sourceMap": false,
-        "module": "CommonJS",
-        "moduleResolution": "Node",
-        "esModuleInterop": true,
-        "skipLibCheck": true,
-        "forceConsistentCasingInFileNames": true,
-        "isolatedModules": true
-    },
-    "exclude": [
-        "node_modules"
-    ],
-    "include": [
-        "./src/*"
-    ],
-    "typeRoots": [
-        "node_modules/@types"
-    ]
-}
