--- common/ethers-ledger-signer/package.json
@@ -1,15 +1,0 @@
-{
-    "name": "@casimir/ethers-ledger-signer",
-    "private": "true",
-    "main": "src/index.ts",
-    "scripts": {
-        "build": "esbuild src/index.ts --bundle --minify --sourcemap --platform=node --target=esnext --outfile=dist/index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "dependencies": {
-        "@ledgerhq/hw-app-eth": "^6.30.4",
-        "@ledgerhq/hw-transport": "^6.27.10",
-        "@ledgerhq/hw-transport-webusb": "^6.27.10",
-        "ethers": "^5.7.2"
-    }
-}
