--- common/ethers-ledger-signer/src/interfaces/EthersLedgerSignerOptions.ts
@@ -1,8 +1,0 @@
-import { ethers } from 'ethers'
-
-export interface EthersLedgerSignerOptions {
-    provider?: ethers.providers.Provider
-    type?: string
-    path?: string
-    baseURL?: string
-}
