--- common/ethers-ledger-signer/tsconfig.json
@@ -1,21 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "ESNext",
-        "strict": true,
-        "preserveConstEnums": true,
-        "noEmit": true,
-        "sourceMap": false,
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "esModuleInterop": true,
-        "skipLibCheck": true,
-        "forceConsistentCasingInFileNames": true,
-        "isolatedModules": true,
-    },
-    "exclude": [
-        "node_modules"
-    ],
-    "include": [
-        "./src/*"
-    ]
-}
