--- common/ethers-trezor-signer/package.json
@@ -1,14 +1,0 @@
-{
-    "name": "@casimir/ethers-trezor-signer",
-    "private": "true",
-    "main": "src/index.ts",
-    "scripts": {
-        "dev": "npx esno src/index.ts",
-        "build": "esbuild src/index.ts --bundle --minify --sourcemap --platform=node --target=esnext --outfile=dist/index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "dependencies": {
-        "@trezor/connect-web": "^9.0.3",
-        "ethers": "^5.7.2"
-    }
-}
