--- common/ethers-trezor-signer/src/interfaces/EthersTrezorSignerOptions.ts
@@ -1,6 +1,0 @@
-import { ethers } from 'ethers'
-
-export interface EthersTrezorSignerOptions {
-    provider?: ethers.providers.Provider
-    path?: string
-}
