--- common/ethers-trezor-signer/src/interfaces/MessageSignature.ts
@@ -1,4 +1,0 @@
-export interface MessageSignature {
-    address: string
-    signature: string
-}
