+++ common/speculos/tsconfig.json
