+++ common/wallets/.gitignore
