+++ common/wallets/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@casimir/bitcoin-ledger-signer",
+  "name": "@casimir/wallets",
   "private": "true",
   "main": "src/index.ts",
   "scripts": {
@@ -8,7 +8,10 @@
   },
   "dependencies": {
     "@ledgerhq/hw-app-btc": "^9.1.2",
+    "@ledgerhq/hw-app-eth": "^6.30.4",
     "@ledgerhq/hw-transport": "^6.27.10",
-    "@ledgerhq/hw-transport-webusb": "^6.27.10"
+    "@ledgerhq/hw-transport-webusb": "^6.27.10",
+    "@trezor/connect-web": "^9.0.3",
+    "ethers": "^5.7.2"
   }
 }
