+++ common/wallets/src/index.ts
@@ -1,0 +1,4 @@
+import { BitcoinLedgerSigner, EthersLedgerSigner } from './providers/ledger'
+import { EthersTrezorSigner } from './providers/trezor'
+
+export { BitcoinLedgerSigner, EthersLedgerSigner, EthersTrezorSigner }
