--- scripts/aws/configure
@@ -1,48 +1,0 @@
-#!/bin/bash
-# Configure process for a specific AWS profile
-#
-# Example:
-#
-#    scripts/aws/configure
-#
-# Further information:
-# See https://docs.aws.amazon.com/
-#
-
-# Get variables from root .env
-export $(xargs < .env)
-
-# Set default profile
-profile="consensus-networks-dev"
-
-if [ ${PROFILE+x} ]; then
-    echo "PROFILE is set to '$PROFILE'"
-    profile=$PROFILE
-else
-    export PROFILE="$profile"
-    echo "PROFILE is not set, using default profile '$PROFILE'"
-fi
-
-# Override project if PROJECT is set in .env
-if [ ${PROJECT+x} ]; then
-    echo "PROJECT is set to '$PROJECT'"
-else
-    export PROJECT="casimir"
-    echo "PROJECT is not set, using default project '$PROJECT'"
-fi
-
-# Override stage if STAGE is set in .env
-if [ ${STAGE+x} ]; then
-    echo "STAGE is set to '$STAGE'"
-else
-    export STAGE="dev"
-    echo "STAGE is not set, using default stage $STAGE"
-fi
-
-export AWS_ACCESS_KEY_ID=$(aws configure get aws_access_key_id --profile $profile)
-export AWS_SECRET_ACCESS_KEY=$(aws configure get aws_secret_access_key --profile $profile)
-
-if [ -z "$AWS_ACCESS_KEY_ID" ] || [ -z "$AWS_SECRET_ACCESS_KEY" ]; then
-    echo "🙈 Could not find AWS credentials for profile '$profile'"
-    exit 1
-fi
