+++ services/users/package.json
@@ -11,7 +11,6 @@
     "cors": "^2.8.5",
     "ethers": "^5.7.2",
     "express": "^4.18.1",
-    "pg": "^8.10.0",
     "supertokens-node": "^13.1.2"
   },
   "devDependencies": {
