+++ services/users/src/providers/users.ts
@@ -1,20 +1,9 @@
 import { userCollection } from '../collections/users'
-import { Postgres } from './postgres'
 import { ProviderString } from '@casimir/types'
 import { User } from '@casimir/types'
 
 
 export default function useUsers() {
-    async function getUser(address: string) {
-        const pg = new Postgres()
-        const text = 'SELECT u.*, json_agg(a.*) AS accounts FROM users u JOIN accounts a ON u.address = a.owner_address WHERE u.address = $1 GROUP BY u.address'
-        const params = [address]
-        const rows = await pg.query(text, params)
-        if (rows.length) {
-            const user: User = rows[0]
-            return user
-        }
-    }
 
     function getMessage (address: string) {
         const user = userCollection.find(user => user.address === address)
@@ -45,5 +34,5 @@ export default function useUsers() {
             * (Number.MAX_SAFE_INTEGER - 1)) + 1).toString()
     }
     
-    return { getUser, getMessage, updateMessage, generateNonce }
+    return { getMessage, updateMessage, generateNonce }
 }
